+++ Engine/Plugins/Runtime/ZoneGraph/Source/ZoneGraphDebug/Private/ZoneGraphTestingActor.cpp
@@ -230,7 +230,7 @@ FPrimitiveSceneProxy* UZoneGraphTestingComponent::CreateSceneProxy()
 					const FMatrix& LocalToWorld = GetLocalToWorld();
 
 					// Taking into account the min and maximum drawing distance
-					const float DistanceSqr = (View->ViewMatrices.GetViewOrigin() - LocalToWorld.GetOrigin()).SizeSquared();
+					const double DistanceSqr = (View->ViewMatrices.GetViewOrigin() - LocalToWorld.GetOrigin()).SizeSquared();
 					if (DistanceSqr < FMath::Square(GetMinDrawDistance()) || DistanceSqr > FMath::Square(GetMaxDrawDistance()))
 					{
 						continue;
